requirements.txt
fastapi
uvicorn
pyngrok
nest-asyncio
python-dotenv
torch
transformers
accelerate
sentencepiece
protobuf